# api/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateLocationSharingRequest:
    properties:
      duration_minutes:
        maximum: 1440
        minimum: 1
        type: integer
      latitude:
        type: number
      longitude:
        type: number
    required:
    - duration_minutes
    - latitude
    - longitude
    type: object
...
  dto.LocationSharingResponse:
    properties:
      expires_at:
        type: string
      id:
        type: string
      share_link:
        type: string
      token:
        type: string
    type: object
...
  dto.PublicLocationResponse:
    properties:
      expires_at:
        type: string
      is_active:
        type: boolean
      last_updated:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      user_name:
        type: string
    type: object
...
  dto.UpdateLocationRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    required:
    - latitude
    - longitude
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: integer
          message: {}
        type: object
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@riskplace.ao
    name: API Support
    url: http://www.riskplace.ao
  description: |-
    This is the API documentation for the Risk Place Angola application.

    ## Environments
    - **Development**: https://risk-place-angola-904a.onrender.com
    - **Local**: http://localhost:8000

    ## WebSocket Integration Guide
    For detailed instructions on integrating WebSocket in mobile applications, please refer to our [WebSocket Notification Guide](https://github.com/risk-place-angola/backend-risk-place/blob/develop/docs/MOBILE_WEBSOCKET_INTEGRATION.md)
  title: Risk Place Angola API
  version: 1.0.0
paths:
...
  /api/v1/location-sharing:
    post:
      consumes:
      - application/json
      description: Create a new location sharing session with expiration time (supports
        both authenticated and anonymous users)
      parameters:
      - description: Device ID for anonymous users
        in: header
        name: X-Device-ID
        type: string
      - description: Location Sharing Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLocationSharingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LocationSharingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create location sharing session
      tags:
      - location-sharing
  /api/v1/location-sharing/{id}:
    delete:
      consumes:
      - application/json
      description: Deactivate an active location sharing session (supports both authenticated
        and anonymous users)
      parameters:
      - description: Device ID for anonymous users
        in: header
        name: X-Device-ID
        type: string
      - description: Location Sharing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stop location sharing
      tags:
      - location-sharing
  /api/v1/location-sharing/{id}/location:
    put:
      consumes:
      - application/json
      description: Update the current coordinates of an active location sharing session
        (supports both authenticated and anonymous users)
      parameters:
      - description: Device ID for anonymous users
        in: header
        name: X-Device-ID
        type: string
      - description: Location Sharing ID
        in: path
        name: id
        required: true
        type: string
      - description: Location Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update shared location coordinates
      tags:
      - location-sharing
...
  /share/{token}:
    get:
      consumes:
      - application/json
      description: Retrieve public location information using share token
      parameters:
      - description: Share Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PublicLocationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: Get shared location by token
      tags:
      - location-sharing
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DeviceID:
    description: Unique device identifier for tracking and analytics.
    in: header
    name: X-Device-ID
    type: apiKey
swagger: "2.0"